requirements.txt
streamlit
pytesseract
Pillow
pdfplumber
deep-translator
pandas
